image/requirements.txt
alpaca-py
numpy
pandas
matplotlib
matplotlib-inline
plotly
seaborn
datetime
scikit-learn
pybase64